+++ angular/src/app/app.module.ts
@@ -1,12 +1,22 @@
-import { NgModule } from '@angular/core';
+import { NgModule, ErrorHandler } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+import { Monitoring } from './data/monitoring.model';
 import { LayoutModule } from './layout/layout.module';
+import { MonitoringService } from './services/monitoring/monitoring.service';
+import { environment } from '../environments/environment';
 
 @NgModule({
+  bootstrap: [AppComponent],
   declarations: [AppComponent],
   imports: [AppRoutingModule, BrowserModule, LayoutModule],
-  bootstrap: [AppComponent],
+  providers: [
+    Monitoring,
+    {
+      provide: ErrorHandler,
+      useClass: environment.production ? MonitoringService : ErrorHandler,
+    },
+  ],
 })
 export class AppModule {}