+++ angular/src/app/services/account/account.service.spec.ts
@@ -29,6 +29,7 @@ describe('AccountService', () => {
           account: 'test',
           booking: null,
           lodging: null,
+          monitoring: null,
         },
         navigation: null,
       };