+++ angular/src/app/services/booking/booking.service.spec.ts
@@ -31,6 +31,7 @@ describe('BookingService', () => {
           account: null,
           booking: 'test',
           lodging: null,
+          monitoring: null
         },
         navigation: null,
       };