+++ angular/src/app/services/lodging/lodging.service.spec.ts
@@ -29,6 +29,7 @@ describe('LodgingService', () => {
           account: null,
           booking: null,
           lodging: 'test',
+          monitoring: null,
         },
         navigation: null,
       };