+++ angular/src/environments/environment.dev.ts
@@ -2,7 +2,7 @@ import * as manifest from '../../package.json';
 
 export const environment = {
   config: 'assets/config.json',
-  name: null,
-  production: false,
+  name: 'DEV',
+  production: true,
   release: `rvtr-app-campsite@${manifest.version}`,
 };