+++ angular/src/environments/environment.prod.ts
@@ -1,3 +1,8 @@
+import * as manifest from '../../package.json';
+
 export const environment = {
-  production: true
+  config: 'assets/config.json',
+  name: 'PROD',
+  production: true,
+  release: `rvtr-app-campsite@${manifest.version}`,
 };